prepare template builder
