+++ .github/workflows/ci.yml
@@ -257,14 +257,14 @@ jobs:
       - uses: actions/checkout@v3
       - name: check for missing metadata
         id: missing-metadata
-        run: cargo run -p build-example-pages -- check-missing
+        run: cargo run -p build-templated-pages -- check-missing examples
       - name: check for missing update
         id: missing-update
-        run: cargo run -p build-example-pages -- update
+        run: cargo run -p build-templated-pages -- update examples
       - name: Check for modified files
         run: |
           echo "if this step fails, run the following command and commit the changed file on your PR."
-          echo " > cargo run -p build-example-pages -- update"
+          echo " > cargo run -p build-templated-pages -- update examples"
           git diff --quiet HEAD --
       - name: Save PR number
         if: ${{ failure() && github.event_name == 'pull_request' }}
