+++ docs-template/EXAMPLE_README.md.tpl
